Move EmailToolTip
