+++ lib/tag-email-tooltip/style.scss
