+++ lib/tag-field/index.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Overlay } from 'react-overlays';
 import isEmailTag from '../utils/is-email-tag';
-import EmailToolTip from '../email-tooltip';
+import EmailToolTip from '../tag-email-tooltip';
 import TagChip from '../components/tag-chip';
 import TagInput from '../tag-input';
 import classNames from 'classnames';
