+++ scss/_components.scss
@@ -13,5 +13,6 @@
 @import 'note-list/style';
 @import 'revision-selector/style';
 @import 'search-field/style';
+@import 'tag-email-tooltip/style';
 @import 'tag-input/style';
 @import 'tag-list/style';
