+++ scss/app.scss
@@ -12,7 +12,6 @@
 @import 'scrollbar';
 @import 'toggle-control';
 
-@import 'tag-editor';
 @import 'placeholders';
 
 // Sections
